refactor: 💡 rename user files to auth for better semantics
renamed userValidationSchemas, userRoutes and userController to auth
prefix for better representation of their authentication functionality
