+++ src/routes/todoRoutes.ts
@@ -1,11 +1,14 @@
 import { Router } from "express";
 import { getTodos, createTodo, updateTodo, deleteTodo } from "../controllers/todoController";
+import { isAuth } from "../middleware/isAuth";
+import { isStudent } from "../middleware/isStudent";
+import { isInstructor } from "../middleware/isInstructor";
 
 const router = Router();
 
-router.get("/", getTodos);
+router.get("/", isAuth, isStudent, getTodos);
 router.post("/", createTodo);
 router.put("/:id", updateTodo);
-router.delete("/:id", deleteTodo);
+router.delete("/:id", isAuth, isInstructor, deleteTodo);
 
 export default router;
